requirements.txt
# ---------- Core Web / Backend ----------
fastapi==0.110.0
uvicorn==0.29.0          # ASGI server for FastAPI
gunicorn==22.0.0         # Optional: production process manager (works with uvicorn workers)

# ---------- Data / Storage ----------
supabase==2.4.0
redis==5.0.3             # redis.asyncio included

# ---------- Trading / Brokers ----------
ccxt==4.4.92             # crypto & FX exchange API wrapper
MetaTrader5==5.0.5120      # MT5 Windows wheel

# ---------- Payments & E-mail / SMS ----------
stripe==8.7.0
coinbase-commerce==1.0.1
flutterwave==1.0.7
sendgrid==6.10.0
twilio==9.0.0
python-jose==3.3.0       # JWT auth helper

# ---------- Config / Metrics ----------
python-dotenv==1.0.1
prometheus-client==0.20.0
apscheduler==3.10.4
cryptography==42.0.8

# ---------- Numerical / TA ----------
numpy==1.26.4
ta==0.10.2               # Technical-analysis indicators

# ---------- Networking ----------
websockets==12.0         # Satisfies realtime (<13, >=11) from supabase

# ---------- Security ----------
pycryptodome==3.20.0

torch @ file:///C:/Users/BLUE%20WAVES/Downloads/torch-2.2.2%2Bcpu-cp310-cp310-win_amd64.whl